requirements.txt
feedparser>=6.0.10
openpyxl>=3.1.2
python-dotenv>=1.0.0
streamlit>=1.32.0
# Optional - only needed for --output gsheet
gspread>=6.0.0
google-auth>=2.27.0
